requirements.txt
Flask==3.1.2
Flask-CORS==6.0.1
Flask-SocketIO==5.5.1
gunicorn==23.0.0
eventlet==0.36.1
PyMySQL==1.1.2
PyJWT==2.10.1
qrcode==8.2
Pillow==11.3.0
requests==2.32.5
firebase-admin==7.1.0
google-generativeai==0.8.5
google-genai==1.47.0
APScheduler==3.11.1
python-socketio==5.14.2
python-engineio==4.12.3
beautifulsoup4==4.12.3
Werkzeug==3.1.3
